Reconfigure project structure
