--- controller/likeController.go
@@ -1,1 +1,0 @@
-package controller
