--- dao/comment.go
@@ -1,1 +1,0 @@
-package dao
