--- service/impl/likeServiceImpl.go
@@ -1,1 +1,0 @@
-package service
