--- service/impl/userServiceImpl.go
@@ -1,1 +1,0 @@
-package service
