--- service/likeService.go
@@ -1,1 +1,0 @@
-package service
