--- service/videoService.go
@@ -1,1 +1,0 @@
-package service
